Return no elements on last round from the round environment in KSP
Due to retaining a Resolver outside KSP's process() a processor could try to find new elements to process in the last round and the resolver, being cached from the last round would return elements that have already been processed. This changes the round environment in KSP to return an empty set when getElementsAnnotatedWith() in the last round. Ideally we shouldn't retain the Resolver and we probably want to null it out before KSP's finish() but that would prevent processors from doing certain work in the last round, such as finding type elements, a breaking behavior change that would require fixing processors first.

Bug: 314783930
Test: XRoundEnvTest.getElementsAnnotatedOnLastRound
Change-Id: I9f5c603ed67f24f9fec8d80b48231a0ae0eb8c7f