Make PagerTest names consistent
SnapPosition.Start, Center, and End are all missing to string and
because we just used that in ParamConfig toString, that meant that
we had an instance like androidx.compose.foundation.gestures.snapping.SnapPosition$Companion$$ExternalSyntheticLambda1@e77d66b
printed in test name. That means that we get a new test name every time
and that causes issues with testing infrastructure.

Test: running PagerTest, names have no hashes
Change-Id: I3f015284ee953b6e4bc7f7f104c664b9a8a9061d