Use firstOrNull for safer lookup of SAM
If an interface simply delegates its declaration to an existing one (via
extends), its corresponding PsiClass does not have SAM in it. Converting
it to UClass is not helpful either in this case. That is, this example
can trigger an exception from `first` for both PsiClass-based lookup and
UClass-based lookup.

Bug: 314048176
Test: unit test added
Change-Id: I0a782ac26436a2765cd3924c89b89f04306d603a