Fix backStackState leak when saveState
We now store an BackStackState only if the entry it is associated with has been actually added to the BackStackMap. This helps prevent leaking backStackStates that never gets restored.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 309559751
Relnote: "Fixed BackStackState leak where multiple saveState calls on a destination would result in multiple states to be saved, but only the first one could be restored."
Change-Id: I598b01c379f61818e96126fb8e3d3ef1f9d084cd