Make SQLiteException an expect class
By making SQLiteException an expect class we can actual type alias it to android.database.SQLException and keep compatibility with existing try-catch that expect to catch Android's exception if a database operation fails.

This change also refactors the driver conformance tests to instead of being a base test class that each driver project uses, it is now an integration project where the drivers are dependencies and each test source set verifies the drivers available in those platforms. This was needed to avoid an issue of the base test being a JVM library where the exception is not type aliased byt used on an Android instrumentation test where it is type aliased. Such moves also means we cleanup the test sources on the driver projects.

Bug: 304297717
Test: ./sqlite/scripts/runConformanceTest
Relnote: "KMP androidx.sqlite.SQLiteException is now type-aliased to android.database.SQLException"
Change-Id: Ia7b8b395229a9f6707f6b3a7ffee9eddd016eca2