Optimize color space conversions
The main thing this change does is introduce a cache of color
space connectors for the built-in color spaces. The cache is
uses an integer key built from the source and destination
color space ids, as well as the render intent. We pre-populate
the cache with our most common conversions (sRGB identity
and sRGB<>OkLab). Combined with various other small changes,
this CL ends up optimizing all conversions.

Importantly this removes a ton of allocations.

On a Pixel 6, before the change:

      226,566   ns      0 allocs ColorBenchmark.colorLerp
      875,048   ns  22545 allocs ColorBenchmark.wideColorLerp

After the change:

      221,855   ns      0 allocs ColorBenchmark.colorLerp
      233,282   ns      0 allocs ColorBenchmark.wideColorLerp

sRGB lerps are 3% faster.
Other lerps are 73% faster and don't allocate after the
first invocation.

Test: ColorSpaceTest, ColorTest
RelNote: Optimized color conversions across color spaces
Change-Id: I1546041eab43b17fca72ea812c107f199c66df8e