Try to reduce screen coordinate test flakiness.
The flakes were failing because the coordinates hadn't updated. This
might be due to lack of synchronization between the system window
manager and the test. AFAIK there's no way to explicitly synchronize
that, so using `waitUntil` to wait for the coordinates to update instead
of just assuming it's ready on idle.

Bug: b/311009092
Test: AndroidComposeViewScreenCoordinatesTest
Relnote: n/a
Change-Id: Ic0924700413cac9215f2ebada439f05288b4d9b6