Removes lazilyCreateIndication parameter
This parameter conflates behavior changes for different use cases, and is a bit misleading. Indication will continue to be created lazily if a null interaction source is provided, and the indication is of type IndicationNodeFactory. Explicitly hoisting an interaction source will disable the lazily creation. Opting into lazy creation for cases where you are hoisting the interaction source but only reading from it will be considered separately in the future.

Bug: b/298048146
Test: tests
Relnote: N/A
Change-Id: I795afdf1c49402def71c87c9de57232444ea57c1