Commonize journal mode setup
For backwards compatibility, it is set in the SupportSQLiteOpenHelper, otherwise the PRAGMA statement is invoked when the connection is opened. Note that AUTOMATIC will only be available in Android as only there we can use 'low-ram device' APIs to figure out mode.

Also add init(config DatabaseConfiguration) to RoomDatabase.native as that will eventually be invoked by the builder.

Bug: 299168035
Bug: 309993778
Test: Existing
Change-Id: I79eb6a54e4f5d02431643964ff70511f36993e67