Consolidate all failure messaging with Fact + Step in FailureMetadata
NOTE TO API COUNCIL: Please ignore this CL, API tracking is only enabled
for Kruth to compare compatibility with Google Truth, and this is an
unpublished library with no intention to publish.

A required change to support the public API that Google Truth uses to
allow users to build custom assertions on custom subjects AND still
allow Truth to provide a way to do soft asserts, including its custom
JUnit rules which allow test assertions against custom truth assertions.

This change consolidates all the possible methods of throwing an
AssertionError into one place from FailureMetadata that is overridable
via custom FailureStrategy, with the failure message generated by a List
of Fact, messagesToPrepend (existing Kruth placeholder) and "Step"s
(internal class supporting builder api for the assertion chain). The
builders around Step allows Truth to differentiate between root subject
and the child subject in more complex assertion chains (e.g., an
IterableSubject generated from the keys of a MapSubject).

Now that we have the same error messaging architecture as Truth, we can
also fix many of the placeholder messages so that they are exactly what
Truth would throw, which is important for compatibility with users who
use the Expect / ExpectFailure API to assert the error messages of
custom subjects, which may inherit or wrap existing subjects provided by
Kruth.

One bad thing is that this introduces a lot of wrapping classes around
messages and considering that Lazy delegation for message creation was
left out, there's a lot we can improve in terms of performance (inline
value class, lazy delegate, etc), but left this for a later CL.

One more thing that was not cleaned up as part of this change is the
numerous fail methods in Subject, which essentially all call into each
other and have varying levels of visibility to support inline reified
and either converting package-private Truth APIs to internal or
protected, since package-private doesn't exist in Kotlin.

This CL also removes data class from FailureMetadata, which is not
recommended anymore for public API due to binary compat guarantees.

Note: This depends on a small part of Guava for lenient String
formatting which was ported into KMP here.

Note: Many of the error messages are still not quite exactly correct,
but fixing them all in one change will make it too difficult to review
this, so left it for a later CL. To be clear this CL does not make
things worse, just doesn't finish fixing all the cases.

Note: Additional tests for Facts + Step depend on ExpectFailure which
has been split into a dependent CL for reviewability.

Test: ANDROIDX_PROJECTS=INFRAROGUE ./gradlew kruth:kruth:allTests -Pandroidx.enabled.kmp.target.platforms=+LINUX
Bug: 317418265
Relnote: "N/A"
Change-Id: Iac9151dad10997d2435545231007f03fa4673785