Fix `XTypeElement#superClass()` for `typealias` superclasses.
This CL fixes the issue in https://github.com/google/dagger/issues/4199.

The bug was due to checking if the `KSDeclaration` was a
`KSClassDeclaration` without taking into consideration that the
declaration could be a `KSTypeAlias`. To fix this, I've replaced the
type aliases in the `KSType` before checking the declaration.

BUG: 317818395
Test: XTypeElementTest.kt
Change-Id: I110f6723fbb9d153a9c6d8dea74b7fbcf9146a74