Fix ValidatingOffsetMapping to skip some cases
ValidatingOffsetMapping checks whether the mapped offset falls in the range of the mapping space. However, it does not validate whether the received offset itself is in a valid range of its space. In this case, IdentityMapping of `VisualTransformation.None` would always fail when wrapped with `ValidatingOffsetMapping` if the received offset is either negative or larger than its space(original or transformed) length.

This change skips the validation step for such cases. Instead the system should be responsible for calling the mapping functions with coerced offsets.

Fix: b/316401857
Test: gradle :compose:foundation:foundation:cAT

Change-Id: Ida4a6ba787b153fd45842ad7aff280eae0742151