Update API files from metalava build ID 11274726
This updates how type parameter lists are written. Previously, "extends java.lang.Object" was only filtered from the end of the string, so it was left in for type parameters not a the end of the list.

The glance ignore file updates are because of a bugfix for the parsing of reified parameters.

Also reenables the androidx_with_metalava target.

Bug: 314790902
Test: presubmit
Relnote: N/A
Change-Id: Iaeb453c1ec9042b1533a61aa4358b24a6c791702