Fix bind failed when StreamSharing is enabled on some devices
The issue is due to that the child crop rect does not have the same
aspect-ratio as the child size after scaling. Not to apply mod 16
computation in calculating child output sizes to fix the issue.

Also, use list instead of set in grouping size by aspect-ratio to
ensure that sizes will be group to 4:3 and 16:9 with priority.

Bug: 319046439
Test: manual test and PreviewLifecycleStatusChangeStressTest
Change-Id: Ic7b6b7dbc431788fbc108c40e8a8729b3248ea5b