Add CaptureConfig ID to the CaptureConfig and the capture pipeline.
Add ID to CaptureConfig and the ID will be passed in every
CameraCaptureCallback methods to enable callers to identify the
source CaptureConfig.

This feature enables CameraX to associate the capture callback to the
source TakePicture request, which will help CameraX to know whether
the capture callback is from aborted request or the active request.

Bug: 319203643
Test: CaptureSessionTest, ProcessingCaptureSessionTest
Change-Id: I4ba056241c8e1d294a1df81122347904cc8c0862