Fix for AndroidView.onUpdate being called after deactivation
As the update calls as posted into a handler, there might be an update call on a node, already being deactivated. We should not proceed with a call in this case, as the node is not active. And when it will be re-activated the update block will be called again.

Test: added a new test in AndroidViewTest
Fixes: 319459482
Change-Id: I90ed53b5c708e1116df1419f6a5ab90fe94bf643