Refactor defaults pattern for Button, TextButton and IconButton
Based on the advice from the API council, this change refactors the
defaults pattern for ButtonColors, TextButtonColors and
IconButtonColors. The idea is to reduce the use of CompositionLocals by
removing the call to `.value` from the default color methods.

To accomplish this, we are adding a new API which will reuse the cached
instances instead of calling the CompositionLocal for each invocation.

Bug: 314818877
Bug: 315289418
Test: Existing performance tests
Relnote: """
Refactor default patterns for ButtonColors, IconButtonColors and
TextButtonColors by creating a cahced instance internally and reducing
the usage of CompositionLocal.
"""

Change-Id: I5f51c87cd5538f78a37318366580883fe8f9d63f