isDefaultIconPreferredAsSingleProvider API Change
We add a bit, isDefaultIconPreferredAsSingleProvider, that, when set to true, tells the to UI prefer rendering
the default credential type icon when the calling provider is the only one. This bit is added across the CredentialEntry API
surface.

Bug: 319723778
Test: Chained in follow up
Relnote: "Extends the credential entry API surface to contain
information on defaulting an icon"

Change-Id: I9fe005a526ccc3347e155146ba8d89852a0a598e