Add lint rule for calling onBackPressed in BackHandler
You should not call onBackPressed on the Activity or
OnBackPressedDispatcher inside of the onBack lambda of a BackHandler or
PredictiveBackHandler. It is an anti-pattern and should be prevented.

This change adds a lint rule to ensure to call this case out at compile
time.

RelNote: "There is now a lint rule to detect calls to `onBackPressed()`
inside of `BackHandler` and `PredictiveBackHandler` callbacks."
Test: added tests
Bug: 287505200

Change-Id: I94d12850523b5b56aa34a5355fdb2228a8ee1236