Suppress RestrictedApiAndroidX in CameraPipe test app
The AndroidX lint checker verifies whether there's any unintended use of
restricted APIs, notably the ones restricted to libraries or library
groups. Unfortunately, it seems that everything under the
integration-tests directory isn't considered to be of the same package
group as the core library code in camera/. As such, we suppress the
warnings coming out of this check in the CameraPipe test app.

Bug: 322236895
Test: Run lint checks for CameraPipe
Change-Id: If4bb5b9aed937d9e76b9461bfa92a32ea12586ce