Change firstUriOrNull behavior
The conventional naming of the function does not match its behavior. In kotlin, firstXOrNull function should iterate all the items and return the first item that is not null. Current `firstUriOrNull` just returns the first uri, it can be null or not. Here we change this behavior to iterate over all the `ClipData.Item`s until a non-null Uri is found, or return null.

Test: AndroidClipboardManagerTest
Change-Id: I59b3fcb59fbf449f3d56409830e13ffd978a243e