Introduce a replace groups to improve detecting changes
When a replace group, startReplaceGroup/endReplaceGroup,
is generated by the compiler it will only generate that
group or a single other group, such as the groups that
surround the then and else blocks of an if statement.
This change avoids triggering the internal differencing
algorithm for these cases.

A corresponding compiler verison is required for these calls
to be generated.

Bug: 322386428
Test: ./gradlew :compose:r:r:tDUT
Relnote: n/a
Change-Id: Iad90f02440e8f7c258ff493afa6cad1c6369be7a