Changes AnchoredDraggable to operate based on small sliding window
Previously, AnchoredDraggable would operate based on `currentValue`, which could be several anchor points away from the current position of the offset. This made exposing APIs reflecting offset fractions tricky. The new behavior is aligned with Pager and subsequently allows us to expose a `currentValueOffsetFraction`.

Relnote: AnchoredDraggable's currentValue will now update when passing through an anchor point. Use settledValue to receive the previous currentValue semantics, only updating when settling at an anchor. The progress is now exposed as a function (requiring a starting and end point) instead of a property.
Test: Existing & refactored tests
Bug: 318707189, 298271489, 294991954
Change-Id: Ibe6e88f172b099e8f1f841722946471e4641f999