Disable reset in multi-swipe when action is committed
In SwipeToReveal, when an action has been committed, the multi-swipe
should not rest the last stat. The reset should only occur if the
actions are not yet committed, which is the state is in Revealing mode.

This change also fixes a bug in the demo where undo action can be
triggered when the SwipeToReveal is collapsing (primary action of
delete is performed).

Fixes: 279896735
Fixes: 279898928
Test: Added new tests and existing tests
Change-Id: Ide059f6818d86e64e98381931c84c1d069de059b