Replace filledButtonColors API with buttonColors
The `filledButtonColors` and `buttonColors` API represent the same
colors and are just duplicated with different names. Based on the API
naming of the component, we are replacing the `filledButtonColors` usage
with `buttonColors` and removing the unnecessary APIs.

Bug: 321861075
Test: Existing tests
Relnote: """
Replace `filledButtonColors` API with `buttonColors` based on the name
of the component and remove the unnecessary API.
"""

Change-Id: I4fe3b136aeb435d35c629e4fa0ab32bd2931b72f