Fix flaky test in CachedPageEventFlowLeakTest
Unable to reproduce the flake organically but suspect that the timeout has to do with polling count from ReferenceQueue in GarbageCollectionTestHelper.
Reduced the timeout from 10 sec to 5 sec so that test doesn't fail from timeout exception, but added assert to ensure GC collected the expected count. This new assert will hopefully help provide better failure message if it flakes again.

Test: ./gradlew paging:paging-common:jvmTest
Fixes: 323086752
Fixes: 322996235
Change-Id: Idfba1bc59be550666c3a1f9be0dd42deb455cbee