Only drop a sub composition if it belongs under a different view
We used to identify a sub-composition by requiring that the
current View id must be one of the ownerViewIds found in Layout
nodes.

It turns out that some sub-compositions do not have a Layout node
that specifies any ownerViewIds. We used to ignore these which led
to missing compose nodes in the Layout Inspector.

Fixes: 322407032
Test: Unit test added.
Change-Id: I8daa8cc5918f774956750e2f17860ad3bed4df82