Fix Text alignment issue when ellipsize, letter spacing and center align is used
This issue exists in Android's TextView and it existed in Compose.

Note that RTL issue still exists in platform.

This applies the fix that Compose has by applying IndentationFixSpan
to the TextView when all these parameters are set.

No need to apply it for non-center alignment as those don't have
issue and there's no diff between applied fix and not applied.
Spannable and SpanText don't have this issue.

Updated existing Material screenshot test cases to use default
center align. We changed them previously to start align because
of this issue, to make them look better, changing it back to this
to test the fix.

Fixes: 319243678
Test: Added
Relnote: "Fixed Text alignment issue when ellipsize, letter spacing and center align are all used on Text."
Change-Id: I716c70d5c35ac599c9bb87eaa1cbd4f2172cbdf9