Replace IdentityArraySet usages with ScatterSet
This change creates a wrapper implementing `Set` around `ScatterSet` that can be unwrapped to operate on underlying storage if needed. The wrapper is passed to the APIs that are exposed publicly and expect `Set` reference.

This change adds two allocations per recomposition cycle: one to pass modified set through apply observer, and second to pass snapshot invalidations to compositions.

Test: runtime tests
Change-Id: I6309be8c13581d67af664cfd5fa4298650c5ec8e