Fixed signature file issue with Collection bulk operation overrides
Previously, due to an error in Psi and a workaround for some, possibly
separate issue, that worked well enough to look ok at a superficial
level, some incorrect signatures were written out for some Collection
bulk operations, i.e. `containsAll`, `removeAll`, `retainAll`.

They should have a signature that looks something like:
```
    boolean retainAll(Collection<? extends X>);
```

Where `X` can be a type parameter or a specific class. Instead what it
did was just write out something like:
```
    boolean retainAll(Collection<E>);
```

No wildcard, and always `E`, even if the actual type was different like
say `T`, or `androidx.compose.ui.text.intl.Locale`. The reason it chose
`E` was because it was given a raw type that it knew should have some
type arguments and it just resolved the class and used its
type parameters as the type arguments. As `Collection` is defined as
`Collection<E>` so `E` was chosen.

More details can be found at https://r.android.com/2953645 in a comment
in the `PsiTypeItem.kt` file.

It was not detected when reading the API signature file as the API
signature parsing code would just treat `E` as if it was an unknown
class in `java.lang` if it did not recognize it as a type parameter.

RelNote: N/A
Bug: 324317136
Test: ./gradlew
Change-Id: I391461ca3af62515c9bc6848c973b674c9329e54