Add MaxTextLength to semantics for a11y
MaxTextLength is used by talkback to announce that the text field has reached its maximum length. Current TextField APIs fully hoist the filtering mechanism, so it's not possible to infer whether a max length filter is applied. Devs will need to explicitly add this info to their semantics.

A follow-up CL will address this requirement for the new BasicTextField(state).

Bug: 170648072
Test: gradle :compose:ui:ui:cAT
Relnote: "Add `maxTextLength` semantics property that should be set on text fields that filter the maximum allowed number of characters."
Change-Id: I24d9f6b2aa6f1e94b67847464314480dd80e8245