Partially migrate PreparedQueryResultBinder to driver API
Specifically InstantPreparedQueryResultBinder, the new CoroutinePreparedQueryResultBinder and the PreparedQueryResultAdapter. This allows for codegen of KMP code for @Query DAO functions with INSERT, UPDATE and DELETE.

One important change is that prepared queries are no longer 'shared' and are prepared, bound and executed on-demand. The effects of this is not measured but should be minimal as Android's cache is deep in the SQLiteSession and not in its statements, meaning holding a reference to an Android statements actually means re-using the binding parameters, but the actual native statement may or may not hit the cache. Statement caching in SQLite KMP (driver APIs) is something to be addressed in the future (b/319653917) so even though this change in theory makes things bit worse it is temporary.

Bug: 299168035
Test: Existing + BaseSimpleQueryTest
Change-Id: Ie5beb8591bbb16fd15fb1e0030281edb8909d892