Optimize allocations in VectorizedMonoSplineKeyframesSpec
Several optimizations to allocations involved in `keyframesWithSpline`
animation, as well as a couple of minor run-time performance
improvements.

For the underlying implementation of the monotic spline (MonoSpline.kt)
Arrays are used in favor of ArrayList and removed an unnecessary
reference copy. These improvements will be noticeable for the first
interpolated frame.

For the vectorized animation implementation, objects are now re-used as
much as possible. This should result in an allocation reduction of at
least 33% when a second animation is triggered. This improvement should be
reflected in the `invalidated_frameToFrame` benchmark.

Additionally, we now avoid an unnecessary linear O(n) search in
MonoSpline.kt, tho this may not reflect on existing benchmark tests
since it was a relatively minor cost in most scenarios.

Bug: n/a
Test: KeyframeSplineAnimationTest passing
Change-Id: I0df17eb31a4edf82ea6ceb96abd210bec6b52358