Generalize PrefetchExecutor.Request interface to prepare for nested prefetch
The goal of this diff is to generalize prefetch requests to operate in terms of `hasWorkToDo` and `execute(availableTime)` instead of directly on `performComposition` and `performMeasure`. This will give us a place to do nested prefetch for nested collections without putting the logic in the PrefetchExecutor itself. It also has the nice side-effect that avg timing tracking is now managed within the list implementation itself, meaning we can more easily update it outside of prefetch and share a single executor across lists.

This change should purely be a refactor which doesn't change what prefetches we do when.

Test: * TrivialListScrollBenchmark/ComplexNestedListsScrollBenchmark: don't see a change
* Verify prefetching in a sample app
* Foundation Unit tests

Relnote: N/A

Change-Id: I3c46826ad769d56a6585d30dd36362a13cb90b59