Count references for nested layers
So we only discardDisplayList() when the layer is not drawn into the parent layer anymore. It allows us to keep the content for disappearance use cases without explicit collection.

Relnote: GraphicsLayer.draw(Canvas) is not a public api anymore. Please use DrawScope.drawLayer(GraphicsLayer) extension function instead to draw the layer.
Test: DrawingPrebuiltGraphicsLayerTest
Change-Id: I7a7c001e58c2f6d4b13562b75a5695bab27aec9d