Move camera operation fields to RestrictedCameraInfo.
Camera operation availabilities are more like camera informations and is
probably better to provide it in RestrictedCameraInfo instead of
RestrictedCameraControl.

Also make RestrictedCameraInfo able to be constructed without a
RestrictedCameraControl, so that it can later be obtained before
binding.

Add utility method SessionProcessor#isOperationSupported, so that it can
be queried directly.

Bug: b/310660777
Test: ./gradlew camera:camera-core:build
Change-Id: Ia1d4806ced2550d5b998211cbbada609deda54a9