Use raw connection in Room code gen utility functions
Update utility functions for code generation to take a lambda function that receives the raw SQLiteConnection. Such change makes it easier to generate code for both Java and Kotlin without having to deal with the connection pool coroutine APIs. Those APIs will still be expose to the user for correctness but internally Room will use the connection more directly. Using a raw connection reduces the number of utility functions created, which are effectively public APIs.

Bug: 299168035
Bug: 322167799
Test: Existing
Change-Id: I155b80a5d5e3da4df6fd3566dcbe34a0bf463472