Detect nested updateData calls
This CL updates DataStoreImpl to throw an exception when updateData is
called inside another updateData block.
Previously, such case would deadlock since the nested call would be
waiting for the parent call to finish which itself is waiting on the nested
call.
This is not a scenerio we can support since we don't know what pending updates
are made in the parent call.
With this change, instead of deadlocking, we'll throw an explict exception
and link to the bug so that the developer can easily detect what is going on.

Fixes: 241760537
Test: SingleProcessDataStoreTest[nestedUpdateCallsShouldntDeadlock, nestedUpdatesInDifferentStores, nestedUpdatesInDifferentStores_fail]
Change-Id: I92db53765d7231f45590b51d8650e41a45c3216b