Ensure non-gesture Systemback has transition in Compose
When using system back, the NavHost `inPredictiveBack` flag flips from
false to true, starting a predictive back gesture, then back to false
which cancels the started gesture and pushes it to completion. This
result in the incoming destination immediately appearing on the screen
instead of transitioning.

We can fix the system back case, by waiting to get progress before
changing the `inPredictiveBack` flag.

There is still the "flick" case that will remain as it is. What we would
need for that is a way to seek the transition to completion from the
current state.

RelNote: "Fixed an issue where NavHost in Compose Navigation failed to
properly transition when using System back without a gesture."
Test: tested in sample app
Bug: 325998468

Change-Id: Iceeae260189dd12af8cb3fe690523646a90f57de