Separate lookahead remeasure and relayout from approach pass
In the past, the invalidation for each node goes through the following
4 stages in sequence:
lookahead measure, approach measure, lookahead place, approach place

The nodes are invalidated from the highest in the tree
to the lowest (i.e. closest to the leaves). It is important to note
that the lookahead place may be cut short as it propagates down
the tree due to placement skipping logic, in the hope that
any lower node that needs lookahead placement will get its
own invalidation call and go through the 4 stages as listed
above. It is therefore possible for a node (lower in the tree)
to be skipped and deferred for lookahead placement, yet the approach
place didn't skip it. The resulting scenario is that placement happens
before lookahead placement.

This CL separates the node invalidation to lookahead invalidation
and approach invalidation, so that lookahead pass can be complete before
the approach during invalidation.

Bug: 325530053
Test: Added
Change-Id: Id7d15add0232b86691c23afbd491673a12202c54