Fix NavArgument equality check for Array args
NavController functions that are based on route rely on `NavDestination#hasMatchingArgs` to match the provided route with a NavDestination.

Previously, the equality check in `hasMatchingArgs` for Array NavTypes erroneously checks for object reference. Now should checks for equality in array content.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 327229511
Relnote: "NavController functions that support routes such as navigate or popBackStack can now properly match routes filled with arguments of Array NavTypes."
Change-Id: Iea8056fe8293e65a80ed53d0e0522967fdb12af1