Remove redundant remember calls
Compose remember calls are used to cache the values against a
composable's key which is retrieved next time it gets called instead of
recomputing the value again. In cases where the computation is trivial
enough (simple comparision), using remember just adds the overhead of
cache retrieval which can be simpler by performing the trivial
computation.

Bug: 294181229
Test: Existing tests
Change-Id: Ide478ea93cd498c6d258c33deeeea4ad2b40cb32