Use node cost for acquiring dynamic node Quota.
Note that this will not impact telemetry node counting as that's being
performed on protos.

RelNote: "Introduce a new getter to DynamicDataNode to retrieve node
cost. The cost is used when acquiring dynamic node quota. Currently,
Nodes with fixed values will have a cost of 0, all the other nodes will
have a cost of 1."
BUG: 316628351
TESTED: Added unit tests.

Change-Id: Ia33e17613a64eb25db850af0d7a5e004638c4941