Add a Coroutine context and scope to Room
As part of the migration to KMP and to coroutines internals, Room will now accept a CoroutineContext in the builder for any coroutine related tasks and queries, such as Flow emissions and in the future invalidation tracker refreshes. A RoomDatabase will only be allowed to be configured with either a coroutine context or executors, but not both.

Backwards compatibility of using the configured Executors in Android will remain, for native and JVM then if no context is provided Dispatchers.IO is used.

Bug: 299168035
Bug: 309996304
Test: Existing
Relnote: "Add a RoomDatabase.Builder setter for a CoroutineContext that will be used to perform Coroutine related queries."
Change-Id: Ib40071a42c18673bbaa2a2ed247919f9f258a2b7