Invert dependency between lifecycle-runtime-compose and compose-ui-ui
This breaks an effective dependency cycle between library groups and will enable the future use of lifecycle-runtime-compose in a multiplatform context without a Compose UI dependency.

Before:
    lifecycle-runtime-compose --> compose-ui-ui --> [lifecycle-runtime & compose-runtime]

After:
    compose-ui-ui --> lifecycle-runtime-compose --> [lifeycle-runtime & compose-runtime]

The definition of the LocalLifecycleOwner composition local previosly lived in compose-ui-ui, which is why lifecycle-runtime-compose took a dependency on it. That composition local provided a LifecycleOwner into the Compose UI-based composition, and lifecycle-runtime-compose provided APIs on top of that LifecycleOwner.

The entirety of lifecycle-runtime-compose's APIs were built on lifecyle-runtime and compose-runtime with the sole exception being the use of the composition local to obtain the LifecycleOwner. By defining the composition local directly within lifecyle-runtime-compose, the dependency can be inverted making compose-ui-ui depend on it (it already depended on the regular runtime). This automatically exposes the Compose-based lifecycle APIs to downstream users of Compose UI and opens up usage of those APIs in compositions other than that provided by Compose UI (e.g., Molecule).

Bug: 328263448
Test: ./gradlew -p compose/ui bOS
Test: ./gradlew -p lifecycle bOS
Relnote: `LocalLifecycleOwner` moved from Compose UI to lifecycle-runtime-compose so that its Compose-based helper APIs can be used outside of Compose UI.
Change-Id: I6c41b92eb6aaab67e7d733dfe3fe0b429b46becf