[PdfViewer] Add MosaicView
A View that can display a very large image (bitmap). Depending on the size at which it has to display, it will either load the whole bitmap and scale it up or down, or operate on a tile map of the bitmap and load the tiles that are visible, or both (use the whole-image bitmap as a backup for missing tiles).

Bug: 327949528
Test: ./gradlew

Change-Id: Ia75e2ad20d05ad2783cc67c3687daa30aa4c892f