Add a useSession and useSessionIn API to CameraPipe
The primary goal of this change is to make it easier to start an
interaction with the camera from a non-suspending function, while
also maintaining ordering, which is non-trivial when using launch
or async to start a new coroutine.

As a result of switching this implementation to use a Mutex
directly, TokenLock is now no longer required and has been removed.

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: Ic8aebcc65fff3e28b0a1090da4541d92f08fd9bd