Commonize MigrationTestHelper
Create and expect declaration of MigrationTestHelper with APIs that use SQLite KMP and will use Room's migration algorithm to validate migrations and schema the same way the Android's version does. The strategy is similar to other commonization, the actual declaration will be thin and will delegate most of the logic into common internal functions that can be used in all platforms. All 3 platforms load schemas differently and thus 3 actual declarations are needed. JVM uses files paths and streams, Android uses instrumentation assets and Native uses file paths (in Strings) and Okio.

For backwards compatibility the Android version of the MigrationTestHelper supports creating an instance of it with and without a SQLiteDriver. When no driver is provided, then the helper is in compatibility mode using a SupportSQLiteOpenHelper. In such configuration the KMP APIs that return SQLiteConnection will throw an exception at runtime. Meanwhile if the helper is configured with a driver then the APIs that return SupportSQLiteDatabase will throw an exception.

Various changes in room-runtime were done to support the test helper. MigrationContainer APIs where made library group restricted, while the connection manager was slightly modified to be usable outside room-runtime.

Bug: 299168035
Bug: 317908685
Test: AutoMigrationTest + Existing
Relnote: "Add MigrationTestHelper APIs that work with SQLite KMP APIs"
Change-Id: Ia648230fb1fea3fc9491d1ee93843db3b74e5f7b