Use unique resource identifier for View based GraphicsLayers
Relnote: "Added resource ids for assist with hiding Views
used as implementation details within build tooling"

Moved the original hide_in_inspector_tag resource id
to the ui:ui module. Renamed the ui:graphics tag to
hide_graphics_layer_in_inspector_tag and updated
ComposeLayoutInspector to search for this flag in
addition to the original flag to determine if a View
should be shown in tooling.

Fixes: 329353088
Test: ran Compose tests
Change-Id: I99531f701d07e95a28aeafd720ec7e999e2254c2