Refactor NavGraph#findNode
findNode for routes need to handle both routePatterns and routeFilled. It used to match routePattern by converting route to an id and matching based on id. But it assumes that the id is always a hashCode of route. This is no longer the case with Safe Args where id can be based on serializer. Now we directly match the string value of routes.

This allows safe args to use route-based APIs that delegates to findNode, such as NavGraph#contains.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Change-Id: I0eb9b9427aa4e2beb45c7bdae0a5939d34bedf14