Add computed composition locals
The change adds the ability to declare a computed composition
local that is a function of existing provided locals.

A computed composition local provided by a parent provider will
calculate the local values provided by children allowing the
child to effectively override a set of related composition locals
by providing one composition local. This can be used, for example,
to specify a base color for which a set of accent colors are computed
by default. Changing the base color will now automatically cause
all the computed colors to change.

Relnote: """
Added the ability to provide a composition local that is computed
based on the values of other composition locals either by default,
using `compositionLocalWithComputedDefault()` or by supplying a
lambda to use to compute the value by using `providesComputed`
instead of `provides`.

The value of a composition local can be obtained by using the
`currentValue` property of the composition local accessible from
within the lambda computing the value. This is used instead of
`current` which is only accessible in `@Composable` functions.
"""

Fixes: 329309693
Test: ./gradelw :compose:r:r:tDUT
Change-Id: Iadbc05f5ce04aeb1b67b53cff265b9017958947a