Fixes not invalidating measure for newly added LayoutAwareModifierNodes
Adding a LayoutAwareModifierNode after layout has already happened (either by recomposing and adding a new node to the chain, or delegating inside an existing node) did not invalidating measure, which meant that the onRemeasured and onPlaced callbacks were never called. For the non-node interface OnRemeasuredModifier (currently used by Modifier.onSizeChanged), BackwardsCompatNode already handles this by invalidating measurement when this is added, so we just need to do the same for LayoutAwareModifierNode as well.

Test: OnSizeChangedTest
Change-Id: I4c595bdd7af0ce01767eb674e45666a020084b73