Remove paddings and insets from scaffold APIs
After deeper investigation and discussion with designers we decided
that it doesn't really make sense to have a scaffold-wide settings
of either paddings or insets. Removes them from the scaffold API
and implementation for now, and clients can implement their own
paddings and insets for each pane. This buys us more time to plan
for a better approach/API to handle paddings and insets.

Relnote: remove paddings and insets from scaffold APIs
Test: existing tests
Bug: 306515112
Change-Id: I786f86697df87c3ed96cdf3cb3065c030057db43